Upgrade mindone transformers to v5.1.0

+++ mindone/transformers/pipelines/image_text_to_text.py
@@ -396,7 +396,7 @@ class ImageTextToTextPipeline(Pipeline):
                 return_tensors=self.framework,
                 tokenize=True,
                 return_dict=True,
-            )
+            ).to(dtype=self.dtype)
             model_inputs["text"] = inputs
             return model_inputs
         # In case we only have text inputs
